requirements.txt
black
botorch
cheetah-accelerator>=0.7.1
flake8
isort
jupyter
lightning
matplotlib
numba
numexpr
numpy
git+https://github.com/ocelot-collab/ocelot.git
pyyaml
retry
SciencePlots
scikit-learn
seaborn
torch
tqdm
xopt>=2.0.0
xsuite
wandb
